+++ packages/react-native-sdk/src/utils/push/android.ts
@@ -1,8 +1,3 @@
-import notifee, {
-  EventType,
-  Event,
-  AndroidCategory,
-} from '@notifee/react-native';
 import { FirebaseMessagingTypes } from '@react-native-firebase/messaging';
 import {
   Call,
@@ -20,6 +15,8 @@ import {
   getFirebaseMessagingLibNoThrow,
   getExpoNotificationsLib,
   getExpoTaskManagerLib,
+  getNotifeeLibThrowIfNotInstalledForPush,
+  NotifeeLib,
 } from './libs';
 import {
   pushAcceptedIncomingCallCId$,
@@ -42,6 +39,14 @@ const ACCEPT_CALL_ACTION_ID = 'accept';
 const DECLINE_CALL_ACTION_ID = 'decline';
 
 type PushConfig = NonNullable<StreamVideoConfig['push']>;
+
+type onBackgroundEventFunctionParams = Parameters<
+  NotifeeLib['default']['onBackgroundEvent']
+>[0];
+
+type Event = Parameters<onBackgroundEventFunctionParams>[0];
+
+//  EventType = NotifeeLib['EventType'];
 
 /** Setup Firebase push message handler **/
 export function setupFirebaseHandlerAndroid(pushConfig: PushConfig) {
@@ -114,6 +119,8 @@ export function setupFirebaseHandlerAndroid(pushConfig: PushConfig) {
   }
 
   // the notification tap handlers are always registered with notifee for both expo and non-expo in android
+  const notifeeLib = getNotifeeLibThrowIfNotInstalledForPush();
+  const notifee = notifeeLib.default;
   notifee.onBackgroundEvent(async (event) => {
     await onNotifeeEvent(event, pushConfig, true);
   });
@@ -208,6 +215,9 @@ const firebaseMessagingOnMessageHandler = async (
       AppState.currentState !== 'active';
     const asForegroundService = canListenToWS();
 
+    const notifeeLib = getNotifeeLibThrowIfNotInstalledForPush();
+    const notifee = notifeeLib.default;
+
     if (asForegroundService) {
       // Listen to call events from WS through fg service
       // note: this will replace the current empty fg service runner
@@ -291,7 +301,7 @@ const firebaseMessagingOnMessageHandler = async (
             },
           },
         ],
-        category: AndroidCategory.CALL,
+        category: notifeeLib.AndroidCategory.CALL,
         fullScreenAction: {
           id: 'stream_ringing_incoming_call',
         },
@@ -316,6 +326,8 @@ const firebaseMessagingOnMessageHandler = async (
       return;
     }
   } else {
+    const notifeeLib = getNotifeeLibThrowIfNotInstalledForPush();
+    const notifee = notifeeLib.default;
     // the other types are call.live_started and call.notification
     const callChannel = pushConfig.android.callChannel;
     const callNotificationTextGetters =
@@ -376,18 +388,24 @@ const onNotifeeEvent = async (
       pushAcceptedIncomingCallCId$.observed &&
       pushRejectedIncomingCallCId$.observed;
 
+    const notifeeLib = getNotifeeLibThrowIfNotInstalledForPush();
+    const notifee = notifeeLib.default;
     // Check if we need to decline the call
     const didPressDecline =
-      type === EventType.ACTION_PRESS &&
+      type === notifeeLib.EventType.ACTION_PRESS &&
       pressAction?.id === DECLINE_CALL_ACTION_ID;
-    const didDismiss = type === EventType.DISMISSED;
+    const didDismiss = type === notifeeLib.EventType.DISMISSED;
     const mustDecline = didPressDecline || didDismiss;
     // Check if we need to accept the call
     const mustAccept =
-      type === EventType.ACTION_PRESS &&
+      type === notifeeLib.EventType.ACTION_PRESS &&
       pressAction?.id === ACCEPT_CALL_ACTION_ID;
 
-    if (mustAccept || mustDecline || type === EventType.ACTION_PRESS) {
+    if (
+      mustAccept ||
+      mustDecline ||
+      type === notifeeLib.EventType.ACTION_PRESS
+    ) {
       clearPushWSEventSubscriptions();
       notifee.stopForegroundService();
     }
@@ -403,16 +421,17 @@ const onNotifeeEvent = async (
       }
       await processCallFromPushInBackground(pushConfig, call_cid, 'decline');
     } else {
-      if (type === EventType.PRESS) {
+      if (type === notifeeLib.EventType.PRESS) {
         pushTappedIncomingCallCId$.next(call_cid);
         // pressed state will be handled by the app with rxjs observers as the app will go to foreground always
-      } else if (isBackground && type === EventType.DELIVERED) {
+      } else if (isBackground && type === notifeeLib.EventType.DELIVERED) {
         pushAndroidBackgroundDeliveredIncomingCallCId$.next(call_cid);
         // background delivered state will be handled by the app with rxjs observers as processing needs to happen only when app is opened
       }
     }
   } else {
-    if (type === EventType.PRESS) {
+    const notifeeLib = getNotifeeLibThrowIfNotInstalledForPush();
+    if (type === notifeeLib.EventType.PRESS) {
       pushTappedIncomingCallCId$.next(call_cid);
       pushConfig.onTapNonRingingCallNotification?.(
         call_cid,